requirements.txt
requests==2.32.3
aiohttp==3.10.5
python-dotenv==1.0.1